html/browsers/browsing-the-web/navigating-across-documents/initial-empty-document/iframe-src-aboutblank-navigate-immediately.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>Navigations on iframe with src set to 'about:blank' immediately after appending to document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<body></body>
<script>
"use strict";
const url1 = "/common/blank.html?1";
const url2 = "/common/blank.html?2";

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through iframe.src.
  iframe.src = url1;
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");

  iframe.src = url2;
  await waitForLoad(t, iframe, url2);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with src");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through setting location.href.
  iframe.contentWindow.location.href = url1;
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");

  iframe.contentWindow.location.href = url2;
  await waitForLoad(t, iframe, url2);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with location.href");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through location.assign().
  iframe.contentWindow.location.assign(url1);
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");

  iframe.contentWindow.location.assign(url2);
  await waitForLoad(t, iframe, url2);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with location.assign");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through window.open().
  iframe.contentWindow.open(url1, "_self");
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");

  iframe.contentWindow.open(url2, "_self");
  await waitForLoad(t, iframe, url2);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with window.open");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through clicking an <a> element.
  const a1 = iframe.contentDocument.createElement("a");
  a1.href = url1;
  iframe.contentDocument.body.appendChild(a1);
  a1.click();
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");

  const a2 = iframe.contentDocument.createElement("a");
  a2.href = url2;
  iframe.contentDocument.body.appendChild(a2);
  a2.click();
  await waitForLoad(t, iframe, url2);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with link click");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank but navigate away from it immediately below.
  const iframe = insertIframeWithAboutBlankSrc(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // Navigate away from the initial empty document through form submission.
  const form1 = iframe.contentDocument.createElement("form");
  form1.action = "/common/blank.html";
  iframe.contentDocument.body.appendChild(form1);
  const input1 = iframe.contentDocument.createElement("input");
  input1.type = "hidden";
  input1.name = "1";
  form1.append(input1);
  form1.submit();
  await waitForLoad(t, iframe, url1 + "=");
  assert_equals(history.length, startingHistoryLength,
    "history.length must not change after normal navigation on initial empty document");

  const form2 = iframe.contentDocument.createElement("form");
  form2.action = "/common/blank.html";
  iframe.contentDocument.body.appendChild(form2);
  const input2 = iframe.contentDocument.createElement("input");
  input2.type = "hidden";
  input2.name = "2";
  form2.append(input2);
  form2.submit();
  await waitForLoad(t, iframe, url2 + "=");
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with form submission");
</script>
